lib: sbi: add support for firmware features extension
This extension allows the software running in supervisor mode to control
the behavior of various features of the SBI. First feature to be support
is SBI_FW_FEATURE_MISALIGNED_DELEG firmware feature which allows to
request delegation of misaligned traps handling to supervisor mode
software. Delegating misaligned traps handling to the OS allows for
finer grain control. For instance when using Linux, calling
prctl(PR_SET_UNALIGN, PR_UNALIGN_SIGBUS) from userspace should change
the behavior of misaligned accesses for the current process in order to
generate a SIGBUS. Thanks to this extension, the misaligned handling can
be delegated to Linux and allow it to handle that behavior properly.

Signed-off-by: Clément Léger <cleger@rivosinc.com>