lib: sbi: sse: assign a domain to SSE event
Now that multiple domains can run at the same time and request
different events, assign them to the requesting domain. Upon injection,
switch to the new domain so that the register content of the previous
domain is saved exactly as it should be restored. When completing, if
the domain was changed, then exit from it after calling the completion
callback.

Signed-off-by: Clément Léger <cleger@rivosinc.com>